Serializable issue in JsonContentFilter fixed and common-cli downgrade
JsonContentFilter implementation had an issue with a reference towards
hamcrest matchers as these are not serializable. The matcher reference
is now marked as being transient and is initialized when the open method
is executed.

Apache Flink provides a common-cli reference which points to version 1.2
whereas the project required 1.3. Unfortunately during startup using the
DefaultParser caused an exception such that the version had to be
downgraded and the DefaultParser needed to be replaced by PosixParser